+++ tests/Feature/Http/HttpTestCase.php
@@ -35,6 +35,8 @@ abstract class HttpTestCase extends TestCase
 
     protected function defineDatabaseMigrations(): void
     {
+        parent::defineDatabaseMigrations();
+
         $this->loadLaravelMigrations();
         $this->loadMigrationsFrom(__DIR__.'/../../../database/migrations');
     }